+++ CHANGELOG.md
@@ -14,6 +14,7 @@
         - ``bindings``: get read-only snapshot of current keybindings
             - modification of the result does not affect actual keybindings
         - ``actions``: get read-only snapshot of current edit actions
+- add builtin ``disown`` command
 
 ### Changed